requirements.txt
PyYAML==6.0.1
pytest==7.4.0
